EC_Customer/requirements.txt
kafka-python
python-dotenv
tkinter 
